Server/chat_service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
email-validator
python-multipart